LayoutTests/http/tests/websocket/tests/hybi/deflate-frame-parameter-expected.txt

Test WebSocket deflate-frame extension.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing query: "max_window_bits=8"
PASS testResult is true
PASS event.data is firstMessage
PASS event.data is secondMessage
onclose() was called.
PASS closeEvent.wasClean is true
Testing query: "no_context_takeover"
PASS testResult is true
PASS event.data is firstMessage
PASS event.data is secondMessage
onclose() was called.
PASS closeEvent.wasClean is true
Testing query: "max_window_bits=8&no_context_takeover"
PASS testResult is true
PASS event.data is firstMessage
PASS event.data is secondMessage
onclose() was called.
PASS closeEvent.wasClean is true
PASS successfullyParsed is true

TEST COMPLETE
